requirements.txt
django
watchdog
django-environ
mysqlclient
django-tinymce
beautifulsoup4
django-mptt
pyclamd
Pillow
phonenumbers
qrcode
django-two-factor-auth
django-otp
django-redis
django-cacheops
paypalrestsdk
stripe

gunicorn